requirements.txt
# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14

# AI / LLM
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Optional: 剧情风格记忆
chromadb>=0.4.22

# HTTP / 环境
httpx>=0.26.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# 工具
python-multipart>=0.0.6
aiofiles>=23.2.1

# 可选：异步任务队列（USE_CELERY=true）
celery[redis]>=5.3.0
redis>=5.0.0
